requirements.txt
flask==3.1.3
groq==0.11.0
flask-cors==6.0.2
gunicorn==25.1.0
waitress==3.0.2
python-dotenv==1.1.1
httpx==0.27.0